Refactor and test the client function
