+++ src/balance_subscriber/__main__.py
@@ -4,9 +4,7 @@ import logging
 import os
 from pathlib import Path
 
-import paho.mqtt.client
-
-import balance_subscriber.callbacks
+from balance_subscriber.client import get_client
 
 logger = logging.getLogger(__name__)
 
@@ -60,26 +58,6 @@ def get_args():
         version=importlib.metadata.version("balance-subscriber"),
     )
     return parser.parse_args()
-
-
-def get_client(
-    topics: set[str], data_dir: Path, encoding: str = "utf-8"
-) -> paho.mqtt.client.Client:
-    if not data_dir:
-        raise ValueError("No data directory specified")
-
-    # Initialise client
-    client = paho.mqtt.client.Client(paho.mqtt.client.CallbackAPIVersion.VERSION2)
-    # https://eclipse.dev/paho/files/paho.mqtt.python/html/index.html#logger
-    client.enable_logger()
-    # Make the topics available to the on_connect callback
-    client.user_data_set(dict(topics=topics, data_dir=data_dir, encoding=encoding))
-
-    # Register callbacks
-    client.on_connect = balance_subscriber.callbacks.on_connect
-    client.on_message = balance_subscriber.callbacks.on_message
-
-    return client
 
 
 def main():
